Conflicts:
	ds-finance-bank-ejb/src/main/java/at/technikum/bic4a16/bi/service/DefaultCompanyService.java
	ds-finance-bank-web/src/main/java/at/technikum/bic4a16/servlets/CompanyServlet.java

+++ ds-finance-bank-common/src/main/java/at/technikum/bic4a16/bi/model/FinancialTransactionRequest.java
@@ -13,6 +13,6 @@ public interface FinancialTransactionRequest extends Serializable{
 
     Customer getCustomer();
 
-    long getNumberOfShares();
+    int getNumberOfShares();
     Action getAction();
 }
